requirements.txt
google-cloud-storage
google-api-core
google-cloud-documentai
